add completed and upcoming appointments and customer dashboard

+++ src/components/NavBar/NavBarBusinessDropdown.js
@@ -1,14 +1,13 @@
 import React, { useContext } from "react";
-import {
-  Link,
-  useHistory,
-} from "react-router-dom";
+import { Link, useHistory } from "react-router-dom";
 
 import "./NavBarBusinessDropdown.css";
 
 import { GlobalContext } from "../../context/GlobalState";
 import { message } from "antd";
 
+import { Url } from "../../constants/ServerUrl";
+import axios from "axios";
 
 function NavBarBusinessDropdown() {
   let history = useHistory();
@@ -16,14 +15,30 @@ function NavBarBusinessDropdown() {
   const contextData = useContext(GlobalContext);
   const name = contextData.loginData.first_name;
   let handleLogout = () => {
-    // to reset global state
-    contextData.setLoginData({})
-    // contextData.setIsLoaded(false)
-    contextData.setRedirectToBusinessInfo(true)
-    localStorage.clear();
-    history.push("/");
-    message.success("You have been successfully logged out.");
+    message.loading({ content: 'Loading...', key:"logout" });
+    axios
+      .get(Url + "/users/logout")
+      .then(function (response) {
+        console.log(response.data);
+        if (response.data.logout===true) {
+          // to reset global state
+          contextData.setLoginData({});
+          // contextData.setIsLoaded(false)
+          contextData.setRedirectToBusinessInfo(true);
+          localStorage.clear();
+          history.push("/");
+          message.success({ content: 'You have been successfully logged out.', key:"logout", duration: 2 });
+  
+        } else {
+          message.error({ content: "Something went wrong. Please try again", key:"logout", duration: 2 });
 
+        }
+      })
+      .catch(function (error) {
+        console.log(error.response);
+        message.error({ content: "Something went wrong. Please try again", key:"logout", duration: 2 });
+
+      });
   };
   return (
     <>
@@ -42,27 +57,24 @@ function NavBarBusinessDropdown() {
 
       <div className="dropdown">
         <Link to="/business/dashboard">
-          {" "}
           <div className="dropdown-item">Dashboard</div>
         </Link>
-        <Link to="/business/appointment">
-          {" "}
-          <div className="dropdown-item">My Appointment</div>
+        <Link to="/business/upcoming-appointments">
+          <div className="dropdown-item">Upcoming Appointments</div>
+        </Link>
+        <Link to="/business/completed-appointments">
+          <div className="dropdown-item">Completed Appointments</div>
         </Link>
         <Link to="/business/services">
-          {" "}
           <div className="dropdown-item">My Services</div>
         </Link>
         <Link to="/business/staff">
-          {" "}
           <div className="dropdown-item">My Staff</div>
         </Link>
         <Link to="/business/account">
-          {" "}
           <div className="dropdown-item">Account</div>
         </Link>
         <Link to="#">
-          {" "}
           <div className="dropdown-item" onClick={handleLogout}>
             Logout
           </div>
