+++ src/components/NavBar/NavBarCustomerDropdown.js
@@ -9,6 +9,8 @@ import "./NavBarCustomerDropdown.css";
 import { GlobalContext } from "../../context/GlobalState";
 import { message } from "antd";
 
+import { Url } from "../../constants/ServerUrl";
+import axios from "axios";
 
 function NavBarCustomerDropdown() {
   let history = useHistory();
@@ -16,14 +18,30 @@ function NavBarCustomerDropdown() {
   const contextData = useContext(GlobalContext);
   const name = contextData.loginData.first_name;
   let handleLogout = () => {
-    // to reset global state
-    contextData.setLoginData({})
-    // contextData.setIsLoaded(false)
-    contextData.setRedirectToBusinessInfo(true)
-    localStorage.clear();
-    history.push("/");
-    message.success("You have been successfully logged out.");
+    message.loading({ content: 'Loading...', key:"logout" });
+    axios
+      .get(Url + "/users/logout")
+      .then(function (response) {
+        console.log(response.data);
+        if (response.data.logout===true) {
+          // to reset global state
+          contextData.setLoginData({});
+          // contextData.setIsLoaded(false)
+          contextData.setRedirectToBusinessInfo(true);
+          localStorage.clear();
+          history.push("/");
+          message.success({ content: 'You have been successfully logged out.', key:"logout", duration: 2 });
+  
+        } else {
+          message.error({ content: "Something went wrong. Please try again", key:"logout", duration: 2 });
 
+        }
+      })
+      .catch(function (error) {
+        console.log(error.response);
+        message.error({ content: "Something went wrong. Please try again", key:"logout", duration: 2 });
+
+      });
   };
   return (
     <>
@@ -42,19 +60,18 @@ function NavBarCustomerDropdown() {
 
       <div className="dropdown">
         <Link to="/customer/dashboard">
-          {" "}
           <div className="dropdown-item">Dashboard</div>
         </Link>
-        <Link to="/customer/appointment">
-          {" "}
-          <div className="dropdown-item">My Appointment</div>
+        <Link to="/customer/upcoming-appointments">
+          <div className="dropdown-item">Upcoming Appointment</div>
+        </Link>
+        <Link to="/customer/completed-appointments">
+          <div className="dropdown-item">Completed Appointment</div>
         </Link>
         <Link to="/customer/account">
-          {" "}
           <div className="dropdown-item">Account</div>
         </Link>
         <Link to="#">
-          {" "}
           <div className="dropdown-item" onClick={handleLogout}>
             Logout
           </div>
